avi_test_framework/requirements.txt
requests==2.31.0
PyYAML==6.0.1
colorama==0.4.6